+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Scenarios without steps can now be excluded from the report by using the new `--exclude-empty-scenarios` report generator option [#151](https://github.com/TNG/JGiven/issues/151)
 * Underlines in parameter names are now replaced with spaces [#147](https://github.com/TNG/JGiven/issues/147)
 * HTML Report: highlight failing cases of a scenario [#150](https://github.com/TNG/JGiven/issues/150)
+* HTML Report: indicate the number of failed cases [#156](https://github.com/TNG/JGiven/issues/156)
 
 # v0.9.3